Move TemplateEntity to template
